Modified for two illuminaitons debug
git-tfs-id: [https://devops.cyberoptics.com/TPC1]$/CyberStitch;C30027

+++ Main/src/logic/ManagedPanelAlignment/ManagedPanelAlignment.cpp
@@ -36,11 +36,14 @@ namespace PanelAlignM {
 	}
 
 	bool ManagedPanelAlignment::AddImage(
-			unsigned int iLayerIndex, 
-			unsigned int iTrigIndex, 
-			unsigned int iCamIndex)
+			int iLayerIndex, 
+			int iTrigIndex, 
+			int iCamIndex)
 	{
-		bool bFlag = _pAligner->AddImage(iLayerIndex, iTrigIndex, iCamIndex);
+		bool bFlag = _pAligner->AddImage(
+			(unsigned int)iLayerIndex, 
+			(unsigned int)iTrigIndex, 
+			(unsigned int)iCamIndex);
 
 		return(bFlag);
 	}
